test: Consolidate localStorage mocks, verify parallel execution (#144)
- Eliminate duplicated localStorage mock implementations from 8 test files
- Create centralized global setup: frontend/__tests__/setup/
  * localStorage-mock.ts: Single source of truth for mock implementation
  * vitest-setup.ts: Global beforeEach/afterEach hooks for test isolation
  * README.md: Documentation for test isolation pattern

- Update frontend/vitest.config.ts: Register setupFiles for global setup

- Refactor all test files to use global setup:
  * frontend/lib/__tests__/apollo-client.test.ts
  * frontend/lib/__tests__/auth-context.test.tsx
  * frontend/__tests__/acceptance-criteria.test.ts
  * frontend/__tests__/integration/auth-errors.test.tsx
  * frontend/__tests__/integration/full-auth-flow.test.tsx
  * frontend/__tests__/integration/multi-user.test.tsx
  * frontend/__tests__/integration/protected-routes.test.tsx
  * frontend/__tests__/integration/security-edge-cases.test.tsx

- Verification (all passing):
  * Sequential: 172 tests ✓
  * Shuffle: 172 tests ✓
  * Parallel: 172 tests ✓

- Benefits:
  * No state leakage between tests
  * Safe parallel execution (--sequence.parallel)
  * Single place to modify mock behavior
  * Clear documentation for future tests
