+++ frontend/__tests__/integration/auth-errors.test.tsx
@@ -6,34 +6,10 @@
  */
 
 import React from 'react';
-import { describe, it, expect, beforeEach, afterEach, vi } from 'vitest';
+import { describe, it, expect, vi } from 'vitest';
 import { render, screen, waitFor } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import '@testing-library/jest-dom/vitest';
-
-// Mock localStorage
-const localStorageMock = (() => {
-  let store: Record<string, string> = {};
-
-  return {
-    getItem: (key: string) => store[key] || null,
-    setItem: (key: string, value: string) => {
-      store[key] = value.toString();
-    },
-    removeItem: (key: string) => {
-      delete store[key];
-    },
-    clear: () => {
-      store = {};
-    },
-    length: 0,
-    key: () => null,
-  };
-})();
-
-Object.defineProperty(globalThis, 'localStorage', {
-  value: localStorageMock,
-});
 
 // Note: Unused GraphQL definitions removed - test uses direct Promise simulation
 // for auth error scenarios without Apollo Client mocking.
@@ -141,15 +117,6 @@ function LoginForm() {
 }
 
 describe('Integration: Auth Error Handling', () => {
-  beforeEach(() => {
-    localStorage.clear();
-  });
-
-  afterEach(() => {
-    localStorage.clear();
-    vi.clearAllMocks();
-  });
-
   describe('Input Validation', () => {
     it('AC#6: Invalid credentials display error message', async () => {
       // Arrange
