+++ frontend/__tests__/integration/protected-routes.test.tsx
@@ -6,36 +6,12 @@
  */
 
 import React from 'react';
-import { describe, it, expect, beforeEach, afterEach, vi } from 'vitest';
+import { describe, it, expect, vi } from 'vitest';
 import { render, screen, waitFor } from '@testing-library/react';
 import { ApolloClient, ApolloError, InMemoryCache, gql } from '@apollo/client';
 import { MockedProvider } from '@apollo/client/testing';
 import { GraphQLError } from 'graphql';
 import '@testing-library/jest-dom/vitest';
-
-// Mock localStorage
-const localStorageMock = (() => {
-  let store: Record<string, string> = {};
-
-  return {
-    getItem: (key: string) => store[key] || null,
-    setItem: (key: string, value: string) => {
-      store[key] = value.toString();
-    },
-    removeItem: (key: string) => {
-      delete store[key];
-    },
-    clear: () => {
-      store = {};
-    },
-    length: 0,
-    key: () => null,
-  };
-})();
-
-Object.defineProperty(globalThis, 'localStorage', {
-  value: localStorageMock,
-});
 
 // Test queries
 const PROTECTED_QUERY = gql`
@@ -92,15 +68,6 @@ function ProtectedComponent() {
 }
 
 describe('Integration: Protected Routes', () => {
-  beforeEach(() => {
-    localStorage.clear();
-  });
-
-  afterEach(() => {
-    localStorage.clear();
-    vi.clearAllMocks();
-  });
-
   describe('Access Control', () => {
     it('AC#5: Dashboard redirects to login when not authenticated', () => {
       // Arrange
