+++ frontend/lib/__tests__/apollo-client.test.ts
@@ -1,38 +1,8 @@
 /* eslint-disable @typescript-eslint/explicit-function-return-type */
-import { describe, it, expect, beforeEach, afterEach, vi } from 'vitest';
+import { describe, it, expect, vi } from 'vitest';
 import { makeClient } from '../apollo-client';
 
-// Mock localStorage
-const localStorageMock = (() => {
-  let store: Record<string, string> = {};
-
-  return {
-    getItem: (key: string) => store[key] || null,
-    setItem: (key: string, value: string) => {
-      store[key] = value.toString();
-    },
-    removeItem: (key: string) => {
-      delete store[key];
-    },
-    clear: () => {
-      store = {};
-    },
-  };
-})();
-
-Object.defineProperty(globalThis, 'localStorage', {
-  value: localStorageMock,
-});
-
 describe('Apollo Client Configuration', () => {
-  beforeEach(() => {
-    localStorage.clear();
-    vi.clearAllMocks();
-  });
-
-  afterEach(() => {
-    localStorage.clear();
-  });
 
   describe('Auth Link - Token Injection', () => {
     it('should inject Bearer token when token exists in localStorage', () => {
